Made minor changes

+++ State coding/Test Code/TestRPM.-Kesav.c
@@ -18,29 +18,29 @@
 #define kP	0.02
 #define kI	0.04
 
-float RPM_Right_Motor;
-float RPM_Left_Motor;
+double RPM_Right_Motor;
+double RPM_Left_Motor;
 
-float RPM_Right_FlyWheel;
-float RPM_Left_FlyWheel;
+double RPM_Right_FlyWheel;
+double RPM_Left_FlyWheel;
 
-float speed_Right = 0;
-float speed_Left = 0;
+double speed_Right = 0;
+double speed_Left = 0;
 
 task main()
 {
-	float pidError_Right = 0;
-	float pidIntegral_Right = 0;
-	float pidValue_Right;
+	double pidError_Right = 0;
+	double pidIntegral_Right = 0;
+	double pidValue_Right;
 
-	float pidError_Left = 0;
-	float pidIntegral_Left = 0;
-	float pidValue_Left;
+	double pidError_Left = 0;
+	double pidIntegral_Left = 0;
+	double pidValue_Left;
 	
-//	float fast = 120,
-//	float med = 90,
-	float slow = 65;
-//	float low = 50;
+//	double fast = 120,
+//	double med = 90,
+	double slow = 65;
+//	double low = 50;
 
 	resetMotorEncoder(leftBack);
 	resetMotorEncoder(rightBack);
@@ -90,7 +90,7 @@ task main()
 
 		motor[rightFront]   = speed_Right;
 		motor[rightBack] = speed_Right;
-		
+
 		motor[leftFront]   = speed_Left;
 		motor[leftBack] = speed_Left;
